add some unit test and main function

+++ src/test/java/AbsTest.java
@@ -1,11 +1,18 @@
+/**
+ * Author: Yun Ni
+ * Student ID: 40179775
+ */
+
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
+
 
 class AbsTest {
 
     @Test
     void absZero() {
+
         assertEquals(0, ExponentFunction.abs(0));
         assertEquals(0, ExponentFunction.abs(-0));
     }
@@ -14,7 +21,7 @@ class AbsTest {
     void absPositive() {
         assertEquals(5, ExponentFunction.abs(5));
         assertEquals(18, ExponentFunction.abs(18));
-        assertEquals(Long.MAX_VALUE, ExponentFunction.abs(Long.MAX_VALUE));
+        assertEquals(Integer.MAX_VALUE, ExponentFunction.abs(Integer.MAX_VALUE));
     }
 
     @Test
@@ -22,6 +29,6 @@ class AbsTest {
         assertEquals(1, ExponentFunction.abs(-1));
         assertEquals(5, ExponentFunction.abs(-5));
         assertEquals(10.89, ExponentFunction.abs(-10.89));
-        assertEquals(Long.MAX_VALUE, ExponentFunction.abs(-Long.MAX_VALUE));
+        assertEquals(Integer.MAX_VALUE, ExponentFunction.abs(-Integer.MAX_VALUE));
     }
 }
